backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Garmin Integration
garminconnect>=0.2.16
garth>=0.4.40
gpxpy==1.6.1

# File Processing
pillow==10.1.0
python-magic==0.4.27

# Data Export
simplekml==1.3.6

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
